requirements.txt
pytest>=9.0.1
selenium>=4.38.0
requests>=2.32.5
allure-pytest>=2.15.3
webdriver-manager>=4.0.2
certifi>=2025.11.12
flake8>=7.3.0
python-dotenv>=1.2.1